[FEATURE] In depth data checking & jQuery usage
The provided request data is not handed over to the `response` or `proxy`
function to allow more in depth decision when mocking responses.

Improve code structure and trigger `onloadstart`, `onload` and `onloadend`
events to make the FakeXMLHttpRequest work underneath the jQuery.ajax()
method.

Get closer to the XMLHttpRequest specification when handling `response`,
`responseText` and `responseXML` properties.